alerts borrar

+++ sportify-app/src/app/pages/my-exercises/my-exercises.component.ts
@@ -3,7 +3,7 @@ import { LoginService } from 'src/app/shared/login.service';
 import { ExerciseService } from 'src/app/shared/exercise-service.service';
 import { Exercise } from 'src/app/models/exercise';
 import { TrainingService } from 'src/app/shared/training-service.service';
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-my-exercises',
@@ -15,49 +15,51 @@ export class MyExercisesComponent implements OnInit {
   public dataBase: object
   public exercise: Exercise
   public options = [
-    {name: 'Selecciona uno', value: 'null'},
-    {name: 'Calentamiento', value:'warmUp'},
-    {name: 'Parte principal', value:'main'},
-    {name: 'Vuelta a la calma', value:'coolDown'}
+    { name: 'Selecciona uno', value: 'null' },
+    { name: 'Calentamiento', value: 'warmUp' },
+    { name: 'Parte principal', value: 'main' },
+    { name: 'Vuelta a la calma', value: 'coolDown' }
   ]
 
-  constructor(private loginService: LoginService, private exService: ExerciseService, private trainingService:TrainingService) { 
-    this.exercise = new Exercise("","","","")
+  constructor(private loginService: LoginService, private exService: ExerciseService, private trainingService: TrainingService) {
+    this.exercise = new Exercise("", "", "", "")
   }
 
-  public getExData(exercise: Exercise){
+  public getExData(exercise: Exercise) {
     this.exercise = exercise
     console.log(this.exercise)
   }
 
-  public postEx(type: HTMLInputElement ,description: HTMLInputElement ,url: HTMLInputElement ,name: HTMLInputElement,){
+  public postEx(type: HTMLInputElement, description: HTMLInputElement, url: HTMLInputElement, name: HTMLInputElement,) {
     let newEx = new Exercise(name.value, description.value, url.value, type.value)
-    newEx.training_id= this.loginService.training_id;
+    newEx.training_id = this.loginService.training_id;
 
-    this.exService.postExercise(newEx).subscribe(data=>{
+    this.exService.postExercise(newEx).subscribe(data => {
       console.log(data)
-      this.exService.getExercise(this.loginService.training_id).subscribe(data =>{
+      this.exService.getExercise(this.loginService.training_id).subscribe(data => {
         console.log(data)
         this.dataBase = data
       })
     })
   }
 
-  public putEx(type: HTMLInputElement ,description: HTMLInputElement ,url: HTMLInputElement ,name: HTMLInputElement ){
+  public putEx(type: HTMLInputElement, description: HTMLInputElement, url: HTMLInputElement, name: HTMLInputElement) {
 
     let newEx = new Exercise(name.value, description.value, url.value, type.value)
+
     newEx.exercise_id= this.exercise.exercise_id;
     console.log(type.value)
 
-    if(name.value===""){
-      newEx.name =this.exercise.name
+    if (name.value === "") {
+      newEx.name = this.exercise.name
     };
-    if(description.value===""){
-      newEx.description=this.exercise.description
+    if (description.value === "") {
+      newEx.description = this.exercise.description
     };
-    if( url.value ===""){
-      newEx.url= this.exercise.url
+    if (url.value === "") {
+      newEx.url = this.exercise.url
     };
+
     if( type.value === null){
       newEx.type = this.exercise.type
     };
@@ -67,25 +69,43 @@ export class MyExercisesComponent implements OnInit {
     this.exService.putExercise(newEx).subscribe(data=>{
       console.log("datos del put")
       console.log(data)
-      this.exService.getExercise(this.loginService.training_id).subscribe(data =>{
+      this.exService.getExercise(this.loginService.training_id).subscribe(data => {
         console.log(data)
         this.dataBase = data
       })
     })
   }
 
-  public deleteEx(ex_id: number){
-    console.log(ex_id)
-    this.exService.deleteExercise(ex_id).subscribe(data=>{
-      this.exService.getExercise(this.loginService.training_id).subscribe(data =>{
-        console.log(data)
-        this.dataBase = data
-      })
+  public deleteEx(ex_id: number) {
+    Swal.fire({
+      title: '¿Estás seguro?',
+      text: "¡No podrás recuperarlo!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#00bfa5',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'borrar'
+    }).then((result) => {
+      if (result.value) {
+        console.log(ex_id)
+        this.exService.deleteExercise(ex_id).subscribe(data => {
+          this.exService.getExercise(this.loginService.training_id).subscribe(data => {
+            console.log(data)
+            this.dataBase = data
+            Swal.fire({
+              title: '¡Eliminado!',
+              text: 'Tu equipo ha sido borrado',
+              icon: 'success',
+              confirmButtonColor: '#00bfa5'
+            })
+          })
+        })
+      }
     })
   }
 
   ngOnInit(): void {
-    this.exService.getExercise(this.loginService.training_id).subscribe(data =>{
+    this.exService.getExercise(this.loginService.training_id).subscribe(data => {
       console.log(data)
       this.dataBase = data
     })
